Updated templates for Live NECTO
