+++ templates/necto/live/project_templates/0_mikrosdk2/template.json
@@ -1,11 +1,13 @@
 {
-    "name": "mikroSDK 2.0 project",
-    "type": "application",
-    "language": "MikroC",
-    "description": "It creates a mikroSDK 2.0, CMake-based application with all dependencies included. The project contains the main.c file with a basic implementation. Outputs: (.hex, .bin)",
-    "project_file": "CMakeLists.txt",
-    "open_in_editor": ["main.c"],
-    "icon": "mikroSDK.png",
-    "files": ["main.c", "CMakeLists.txt"],
-    "legacy" : false
+  "name": "Standard project",
+  "type": "mikrosdk",
+  "language": "C",
+  "description": "Create a standard project in C language. Suitable for most common development.",
+  "project_file": "CMakeLists.txt",
+  "open_in_editor": ["main.c"],
+  "icon": "fa-code",
+  "files": ["CMakeLists.txt", "main.c"],
+  "files_included": "**CMakeLists.txt:** Configuration file to build your project.\n**main.c:** Source file with empty main function.",
+  "outputs": "**.hex:** Hexadecimal binary for MCU flashing.\n**.bin:** Raw binary for various deployment options.",
+  "requirements": "No specific requirements."
 }
