+++ templates/necto/live/project_templates/3_designer/CMakeLists.txt
