+++ templates/necto/live/project_templates/3_designer/main_screen.c
