+++ templates/necto/live/project_templates/3_designer/resource.mres
