--- templates/necto/live/project_templates/3_library_sdk/template.json
@@ -1,10 +1,0 @@
-{
-    "name": "mikroSDK Library project",
-    "type": "library",
-    "language": "MikroC",
-    "description": "Creates a static library. Outputs: archive (.a)",
-    "project_file": "CMakeLists.txt",
-    "open_in_editor": ["lib.c"],
-    "icon": "library_normal.mpng",
-    "files": ["CMakeLists.txt", "lib.c", "lib.h"]
-}
