+++ templates/necto/live/project_templates/4_application_with_library/main.c
@@ -1,26 +1,33 @@
 /**
  * @file main.c
- * @brief Main function for ${PROJECT_NAME} application.
+ * @brief Main source file for the ${PROJECT_NAME} application.
+ *
+ * This is a basic template for a standalone embedded application.
+ * It provides a minimal structure for user code and MCU initialization.
  */
 
-/**
- * Any initialization code needed for MCU to function properly.
- * Do not remove this line or clock might not be set correctly.
- */
 #ifdef PREINIT_SUPPORTED
 #include "preinit.h"
 #endif
 
+/**
+ * @brief Application entry point.
+ *
+ * Initializes the MCU and starts the main execution loop.
+ * Do not remove the preinit() call — it ensures proper clock configuration.
+ */
 int main(void)
 {
-    /* Do not remove this line or clock might not be set correctly. */
-    #ifdef PREINIT_SUPPORTED
+    /* Do not remove this line — it ensures correct MCU initialization. */
+#ifdef PREINIT_SUPPORTED
     preinit();
-    #endif
+#endif
 
-    /* Replace with your application code */
+    // Add your application initialization code here.    
+
     while (1)
     {
+        // Place your program logic here.
     }
 
     return 0;
