--- templates/necto/live/project_templates/4_library_with_example/CMakeLists.txt
@@ -1,50 +1,0 @@
-cmake_minimum_required(VERSION 3.19)
-
-if (${TOOLCHAIN_LANGUAGE} MATCHES "MikroC")
-    project(${PROJECT_NAME}_example LANGUAGES MikroC)
-else()
-    project(${PROJECT_NAME}_example LANGUAGES C ASM)
-endif()
-
-include(mikroeUtils)
-
-set(PROJECT_TYPE "${PROJECT_TYPE_VALUE}" CACHE STRING "" FORCE)
-
-add_executable(${PROJECT_NAME}_example main.c)
-############################ ${PROJECT_NAME} GENERATED CODE START ###########################
-############################ ${PROJECT_NAME} GENERATED CODE END ###########################
-
-find_package(MikroC.Core REQUIRED)
-target_link_libraries(${PROJECT_NAME}_example
-    PUBLIC
-        MikroC.Core
-        ${PROJECT_NAME}
-)
-
-if (${MIKROSDK_TYPE} STREQUAL "legacy")
-    find_package(MikroC.System REQUIRED)
-    target_link_libraries(${PROJECT_NAME}_example
-        PUBLIC
-            MikroC.System
-            ${PROJECT_NAME}
-    )
-endif()
-
-if(COMPILER_REVISION)
-    if(${COMPILER_REVISION} VERSION_GREATER_EQUAL "1.0")
-        if(${TOOLCHAIN_LANGUAGE} STREQUAL "GNU")
-            if(EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/.meproject)
-                fetch_startup_name(${CMAKE_CURRENT_SOURCE_DIR}/.meproject startupName)
-                if(startupName)
-                    add_startup_all_targets(. ${CMAKE_CURRENT_SOURCE_DIR} ${startupName})
-                endif()
-                fetch_linker_name(${CMAKE_CURRENT_SOURCE_DIR}/.meproject linkerName)
-                if(linkerName)
-                    add_ld_all_targets(. ${CMAKE_CURRENT_SOURCE_DIR} ${linkerName})
-                endif()
-            endif()
-        endif()
-    endif()
-endif()
-
-add_subdirectory(lib)
