--- templates/necto/live/project_templates/4_library_with_example/lib/lib.c
@@ -1,6 +1,0 @@
-/**
- * @file lib.c
- * @brief ${PROJECT_NAME} library.
- */
-
-/* User code here */
