--- templates/necto/live/project_templates/4_library_with_example/lib/lib.h
@@ -1,6 +1,0 @@
-/**
- * @file lib.h
- * @brief ${PROJECT_NAME} library.
- */
-
-/* User code here */
