--- templates/necto/live/project_templates/5_dual_core/core_0_dir_name/main.c
@@ -1,27 +1,0 @@
-/**
- * @file main.c
- * @brief Main function for the primary core of the dual-core application.
- */
-
-/**
- * Any initialization code needed for MCU to function properly.
- * Do not remove this line or clock might not be set correctly.
- */
-
-#ifdef PREINIT_SUPPORTED
-#include "preinit.h"
-#endif
-
-int main(void)
-{
-    /* Do not remove this line or clock might not be set correctly. */
-    #ifdef PREINIT_SUPPORTED
-    preinit();
-    #endif
-
-    while (1)
-    {
-    }
-
-    return 0;
-}
