+++ templates/necto/live/project_templates/5_library_sdk/lib.h
