+++ templates/necto/live/toolchain_templates/mikroCToolchain.cmake
@@ -4,7 +4,9 @@ set(COMPILER_FLAGS %COMPILER_FLAG_VALUES% "-DBG")
 set(LINKER_FLAGS %LINKER_FLAG_VALUES% "-DBG")
 
 string(TOLOWER ${CMAKE_BUILD_TYPE} build_type)
-if (build_type STREQUAL debug)
+set(DEBUG_SUPPORTED %IS_DEBUG_SUPPORTED%)
+
+if (DEBUG_SUPPORTED AND build_type STREQUAL "debug")
     list(APPEND COMPILER_FLAGS "-UICD")
     list(APPEND LINKER_FLAGS "-UICD")
 endif()
